Support for snapshot versioning and publish under custom tag (#359)
* initial implementation of snapshot release

* add logic to release changelog file too

* use unique hash for versioning

* lint fix - remove unused vars

* add error message when user runs snapshot in pre state

* revert unused code from pre command

* use hash cache

* revert adding logger to git

* update changelog for snapshot release support

* 1: [Version command] Move caching of snapshot version to the assemble-release-plan's main function
2: [Version command] Add condition to throw error if snapshot version is made while changeset is in pre-state
3: [Publish command] Update the condision to capture preState.mode and updated the error message when tag is passed in pre state

* split changessets and updated the description

* Update .changeset/violet-gorillas-march.md


* Update .changeset/dry-rivers-train.md

+++ packages/apply-release-plan/src/index.ts
@@ -39,7 +39,8 @@ async function getCommitThatAddsChangeset(changesetId: string, cwd: string) {
 export default async function applyReleasePlan(
   releasePlan: ReleasePlan,
   packages: Packages,
-  config: Config = defaultConfig
+  config: Config = defaultConfig,
+  snapshot?: string | boolean
 ) {
   let cwd = packages.root.dir;
 
@@ -73,7 +74,7 @@ export default async function applyReleasePlan(
     cwd
   );
 
-  if (releasePlan.preState !== undefined) {
+  if (releasePlan.preState !== undefined && snapshot === undefined) {
     if (releasePlan.preState.mode === "exit") {
       await fs.remove(path.join(cwd, ".changeset", "pre.json"));
     } else {
